requirements.txt
streamlit
google-generativeai
groq
python-dotenv
pandas
requests
